requirements.txt
pandas
requests
pymongo
mysql-connector-python
python-dotenv